+++ CHANGES.md
@@ -2,6 +2,8 @@ v0.2.11 (dev)
 =============
 + more informative error message on bad sample name (#53)
 + allow setting SOMALIER_AB_HOM_CUTOFF to change which calls are considered hom-ref (#56)
++ adjust (fix) relatedness calculation which was off when the number of shared sites was low (#55).
+  many thanks to @fgvieira who found and diagnosed this problem.
 
 v0.2.10
 =======